user login, media, and modify

+++ app/Http/Controllers/Dashboard/MediaController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
+use App\Models\Category;
 use App\Models\Media;
+use App\Models\Product;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -24,8 +26,15 @@ class MediaController extends Controller
      */
     public function store(Request $request)
     {
+        $request->validate([
+            'imageFile' => 'required|array|min:1',
+            'imageFile.*' => 'image|mimes:jpeg,png,jpg,gif|max:2048',
+        ],[
+            'imageFile.*' => 'يجب تحميل صورة فقط باصدار jpeg,png,jpg,gif',
+        ]);
         if($request->hasFile('imageFile')){
             foreach ($request->file('imageFile') as $imageFile){
+
                 $imageName = time() . ' - ' . $imageFile->getClientOriginalName() . '.' . $imageFile->extension();
                 $imagePath = $imageFile->storeAs('images', $imageName, 'public');
                 $mediaData = [
@@ -52,12 +61,29 @@ class MediaController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
         $image  = Media::findOrFail($id);
+        // الفحص في جدول الأصناف والمنتجات اذا وجدت الصورة
+        $categoryImage = Category::where('image', $image->path)->first();
+        $productImage = Product::where('image', $image->path)->first();
+        if($categoryImage != null || $productImage != null){
+            $confirmation_deletion = $request->confirmation_deletion;
+            if($confirmation_deletion == null){
+                return response()->json(['error' => 'لا يمكن حذف هذه الصورة بسبب تحميلها لاحدى المنتجات والأصناف', 'confirmation_deletion' => false], 400);
+            }
+        }
         $image_old = $image->path;
         if($image_old != null){
             Storage::delete('public/'.$image_old);
+        }
+        if($categoryImage != null){
+            $categoryImage->image = null;
+            $categoryImage->save();
+        }
+        if($productImage != null){
+            $productImage->image = null;
+            $productImage->save();
         }
         $image->delete();
         return response()->json(['success' => 'تم حذف الصورة بنجاح']);
